requirements.txt
Django==5.1.7
python-dateutil==2.9.0.post0
gunicorn==23.0.0
psycopg2==2.9.10
coverage==7.8.0
